+++ common/src/main/java/com/telepathicgrunt/the_bumblezone/modcompat/ResourcefulBeesCompat.java
@@ -3,12 +3,15 @@ package com.telepathicgrunt.the_bumblezone.modcompat;
 import com.telepathicgrunt.the_bumblezone.Bumblezone;
 import com.telepathicgrunt.the_bumblezone.configs.BzModCompatibilityConfigs;
 import com.telepathicgrunt.the_bumblezone.events.entity.EntitySpawnEvent;
+import com.telepathicgrunt.the_bumblezone.events.lifecycle.AddBuiltinDataPacks;
+import com.telepathicgrunt.the_bumblezone.events.lifecycle.AddBuiltinResourcePacks;
 import com.telepathicgrunt.the_bumblezone.utils.GeneralUtils;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.HolderSet;
 import net.minecraft.core.Registry;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.core.registries.Registries;
+import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.tags.TagKey;
 import net.minecraft.util.RandomSource;
@@ -44,6 +47,14 @@ public class ResourcefulBeesCompat implements ModCompat {
         if (BEE_JAR.isPresent() && BzModCompatibilityConfigs.allowResourcefulBeesBeeJarRevivingEmptyBroodBlock) {
             setupDispenserCompat(BEE_JAR.get()); // adds compatibility with bee jars in dispensers
         }
+
+        Bumblezone.MOD_COMPAT_DATAPACKS.add(addBuiltinDataPacks ->
+            addBuiltinDataPacks.add(
+                new ResourceLocation(Bumblezone.MODID, "resourcefulbees_compat"),
+                Component.literal("Bumblezone - Resourceful Bees Compat"),
+                AddBuiltinDataPacks.PackMode.FORCE_ENABLED
+            )
+        );
 
         // Keep at end so it is only set to true if no exceptions was thrown during setup
         ModChecker.resourcefulBeesPresent = true;