+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ### **(V.7.5.1 Changes) (1.20.6 Minecraft)**
 
+##### Entities:
+Fixed breeding Honey Slime not counting towards the Bee Queen advancement for breeding them.
+
 ##### Dimension:
 Added `the_bumblezone:fog_adjusting_effects` mob effect tag to allow specifying what mob effect changes the fog in Bumblezone.
  By default, this has Blindness and Darkness. This new addition now fixes those two to properly set the fog distance in Bumblezone dimension.